requirements.txt
python-dotenv
yandex-music
python-telegram-bot
